+++ internal/hosting/bitbucketcloud/connector.go
@@ -1,4 +1,4 @@
-package bitbucket
+package bitbucketcloud
 
 import (
 	"errors"
