+++ internal/hosting/bitbucketcloud/connector_test.go
@@ -1,4 +1,4 @@
-package bitbucket_test
+package bitbucketcloud_test
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 	"github.com/git-town/git-town/v16/internal/config/configdomain"
 	"github.com/git-town/git-town/v16/internal/git/gitdomain"
 	"github.com/git-town/git-town/v16/internal/git/giturl"
-	"github.com/git-town/git-town/v16/internal/hosting/bitbucket"
+	"github.com/git-town/git-town/v16/internal/hosting/bitbucketcloud"
 	"github.com/git-town/git-town/v16/internal/hosting/hostingdomain"
 	. "github.com/git-town/git-town/v16/pkg/prelude"
 	"github.com/shoenig/test/must"
@@ -22,7 +22,7 @@ func TestBitbucketConnector(t *testing.T) {
 			t.Parallel()
 			url, has := giturl.Parse("username@bitbucket.org:git-town/docs.git").Get()
 			must.True(t, has)
-			have := bitbucket.NewConnector(bitbucket.NewConnectorArgs{
+			have := bitbucketcloud.NewConnector(bitbucketcloud.NewConnectorArgs{
 				HostingPlatform: None[configdomain.HostingPlatform](),
 				RemoteURL:       url,
 			})
@@ -38,7 +38,7 @@ func TestBitbucketConnector(t *testing.T) {
 			t.Parallel()
 			url, has := giturl.Parse("git@custom-url.com:git-town/docs.git").Get()
 			must.True(t, has)
-			have := bitbucket.NewConnector(bitbucket.NewConnectorArgs{
+			have := bitbucketcloud.NewConnector(bitbucketcloud.NewConnectorArgs{
 				HostingPlatform: Some(configdomain.HostingPlatformBitbucket),
 				RemoteURL:       url,
 			})
@@ -55,7 +55,7 @@ func TestBitbucketConnector(t *testing.T) {
 		t.Parallel()
 		url, has := giturl.Parse("username@bitbucket.org:org/repo.git").Get()
 		must.True(t, has)
-		connector := bitbucket.NewConnector(bitbucket.NewConnectorArgs{
+		connector := bitbucketcloud.NewConnector(bitbucketcloud.NewConnectorArgs{
 			HostingPlatform: None[configdomain.HostingPlatform](),
 			RemoteURL:       url,
 		})
