+++ internal/hosting/bitbucketcloud/core.go
@@ -1,2 +1,2 @@
 // Package bitbucket provides the code hosting connector for BitBucket.
-package bitbucket
+package bitbucketcloud
